@@ -31,16 +31,15 @@
     "url": "https://github.com/puppeteer/puppeteer/tree/main/packages/puppeteer-core"
   },
   "engines": {
-    "node": ">=14.1.0"
+    "node": ">=14.14.0"
   },
   "scripts": {
-    "build:third_party": "wireit",
+    "build:docs": "wireit",
     "build:tsc": "wireit",
     "build:types": "wireit",
     "build": "wireit",
     "check": "tsx tools/ensure-correct-devtools-protocol-package",
-    "clean": "tsc -b --clean && rimraf lib src/generated",
-    "format:types": "wireit",
+    "clean": "tsc -b --clean && rm -rf lib src/generated",
     "generate:package-json": "wireit",
     "generate:sources": "wireit",
     "prepack": "wireit"
